More precise int lengths

+++ src/notation.rs
@@ -56,7 +56,7 @@ pub struct PiecesPlacement {
 ///    and is incremented after Black's move.
 pub fn parse_fen
     (s: &str)
-     -> Result<(PiecesPlacement, Color, CastlingRights, Option<Square>, u32, u32), ParseError> {
+     -> Result<(PiecesPlacement, Color, CastlingRights, Option<Square>, u8, u16), ParseError> {
 
     let fileds: Vec<_> = s.split_whitespace().collect();
     if fileds.len() == 6 {
@@ -64,8 +64,8 @@ pub fn parse_fen
             try!(parse_fen_active_color(fileds[1])),
             try!(parse_fen_castling_rights(fileds[2])),
             try!(parse_fen_enpassant_square(fileds[3])),
-            try!(fileds[4].parse::<u32>().map_err(|_| ParseError)),
-            try!(fileds[5].parse::<u32>().map_err(|_| ParseError))))
+            try!(fileds[4].parse::<u8>().map_err(|_| ParseError)),
+            try!(fileds[5].parse::<u16>().map_err(|_| ParseError))))
     } else {
         Err(ParseError)
     }
